# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Space Case Cruises API
  version: 0.2.0
  description: |
    Admin-first community platform API for Space Case Cruises.
    Auth in local/dev uses one of these headers:
    - x-user-email
    - x-user-id
servers:
  - url: http://localhost:4000
security:
  - UserEmailAuth: []
  - UserIdAuth: []
components:
  securitySchemes:
    UserEmailAuth:
      type: apiKey
      in: header
      name: x-user-email
    UserIdAuth:
      type: apiKey
      in: header
      name: x-user-id
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Role:
      type: string
      enum: [user, admin]
...
paths:
...
  /api/v1/admin/cadets/{id}:
    patch:
      tags: [Admin Cadre]
      summary: Admin update cadet role/extension/disabled
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                cadet_extension:
                  type: string
                role:
                  $ref: '#/components/schemas/Role'
                is_disabled:
                  type: boolean
      responses:
        '200':
          description: Updated cadet
        '403':
          description: Forbidden
...
  /api/v1/cruise-subgroups/{id}:
    get:
      tags: [Commitments]
      summary: Get subgroup detail + commitment count
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - in: query
          name: include_committed_cadets
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        '200':
          description: Subgroup detail
  /api/v1/commitments/toggle:
    post:
      tags: [Commitments]
      summary: Transition commitment lifecycle for current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [cruise_subgroup_id, action]
              properties:
                cruise_subgroup_id:
                  type: string
                  format: uuid
                action:
                  type: string
                  enum: [commit, withdraw, recommit]
      responses:
        '200':
          description: Updated commitment
        '409':
          description: Invalid transition or completed commitment
...
  /api/v1/admin/badges:
    post:
      tags: [Badges]
      summary: Create badge (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                  nullable: true
                icon_url:
                  type: string
                  format: uri
                  nullable: true
                cruise_id:
                  type: string
                  format: uuid
                  nullable: true
      responses:
        '201':
          description: Created badge
  /api/v1/admin/badges/{id}:
    patch:
      tags: [Badges]
      summary: Update badge (admin)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  type: string
                description:
                  type: string
                  nullable: true
                icon_url:
                  type: string
                  format: uri
                  nullable: true
                cruise_id:
                  type: string
                  format: uuid
                  nullable: true
      responses:
        '200':
          description: Updated badge
  /api/v1/admin/badge-assignments:
    post:
      tags: [Badges]
      summary: Assign badge to cadet (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [user_id, badge_id]
              properties:
                user_id:
                  type: string
                  format: uuid
                badge_id:
                  type: string
                  format: uuid
                reason:
                  type: string
                  nullable: true
    get:
      tags: [Badges]
      summary: List badge assignments (admin)
      responses:
        '200':
          description: Assignment list
  /api/v1/admin/badge-assignments/{id}/revoke:
    patch:
      tags: [Badges]
      summary: Revoke badge assignment (admin)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Revoked assignment